Finalização da implementação do item: - Dado o nome de um produto, retornar um objeto com nome, fabricante, categoria, sua avaliação média e uma lista contendo todos os preços e as lojas em que o objeto é vendido.

+++ CoruscantMarketplace.Core.Models/ResumoProduto.cs
@@ -32,7 +32,7 @@ namespace CoruscantMarketplace.Core.Models
         /// Média da recomendação do produto
         /// </summary>
         [JsonProperty("avaliacao")]
-        public int Avaliacao { get; set; }
+        public double Avaliacao { get; set; }
 
         /// <summary>
         /// Coleção com o nome da loja e o preço
